Move formatting functions to a module walt.common.formatting
