+++ server/walt/server/postgres.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 from walt.server.const import WALT_DBNAME, WALT_DBUSER
-from walt.server.tools import columnate
+from walt.common.formatting import columnate
 import psycopg2, shlex, uuid
 from psycopg2.extras import NamedTupleCursor
 from subprocess import Popen, PIPE
