+++ server/walt/server/threads/main/devices/manager.py
@@ -2,8 +2,9 @@
 from walt.server.threads.main.network.tools import set_static_ip_on_switch, \
                             ip_in_walt_network, get_walt_subnet, get_server_ip
 from walt.server.threads.main.network.netsetup import NetSetup
-from walt.common.tools import format_sentence, get_mac_address
-from walt.server.tools import format_paragraph, to_named_tuple, merge_named_tuples
+from walt.common.tools import get_mac_address
+from walt.common.formatting import format_paragraph, format_sentence
+from walt.server.tools import to_named_tuple, merge_named_tuples
 from walt.server import const
 import re, json
